docs: refactor scripts
