+++ docs/static/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>jsonx</title>
-    <link rel="stylesheet" href="./play.css" />
+    <link rel="stylesheet" href="./lib/playground/playground.css" />
   </head>
   <body>
     <nav>
