+++ docs/static/lib/playground/index.js
@@ -1,0 +1,1 @@
+export * from "./playground.js";
